server/crawler/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
filelock==3.12.0
Flask==2.2.3
Flask-Cors==3.0.10
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.5.0
gunicorn==20.1.0
huggingface-hub==0.14.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nltk==3.8.1
numpy==1.24.2
openai==0.27.1
packaging==23.0
pdfminer==20191125
Pillow==9.5.0
pycparser==2.21
pycryptodome==3.17
PyMySQL==1.0.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.5.5
requests==2.28.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post1
soupsieve==2.4
sympy==1.12
textblob==0.17.1
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.29.2
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
yarl==1.8.2
